Built demo of how I should plot the comparison sections.
Key thing to remember with GGPLOT. When I want to put a profile on every facet, I need to add another column that is the faceting variable. Then I take whatever trajectory I want to plot and include it in EVERY instance of the faceting variable. Not efficient from a storage POV, but oh man, when I remember it, it makes putting out good looking plots quickly very easy.